requirements.txt
Flask==2.0.1
requests==2.28.1
python-dotenv==0.20.0
pyyaml==6.0 